Fix the JSON body since Request 2.50.0 update
with 2.48, I received on my server the following JSON: "{"login":"login","password":"password"}" which is basically the JSON.stringify() output of the above object.
with 2.50, I received on my server the following JSON: 
""{\"login\":\"login\",\"password\":\"password\"}"" which seems to be a JSON stringified twice...

If I patch this line by removing the JSON.stringify() call and passing directly the data object, it works again.

Lastly, the Request doc states:

body - entity body for PATCH, POST and PUT requests. Must be a Buffer or String, unless json is true. If json is true, then body must be a JSON-serializable object.
This is why I got the idea of removing the JSON.stringify() call since Request is expecting an object.