Removed JSON.stringify to prevent double stringify
requestjs applies JSON.stringify to the request body when json option is true. Stringifying from this module makes the requestjs to send request body as "{}" instead of {} making the server throw "invalid json" when bodyparser module is used.

I just removed JSON.stringify in this module to fix the issue.