Fix for jsonContains/Types nested object comparisons.
In _jsonContains and _jsonContainsTypes the arguments.callee.caller.name
check returns false when the functions recurse when evaluating callbacks
within nested objects. This fix will evaluate callbacks in nested objects
when the callee.name is the same as the caller.name.

This fix also evaluates callbacks within arrays by checking for the
"array" type.